+++ R/utils.R
@@ -42,7 +42,13 @@ geomean <- function(x, na.rm = TRUE){
 #' @keywords internal
 #' @noRd
 checkNames <- function(x, namesVector){
+
   goodNames <- colnames(x)
+
+  # remove columns specified as empty strings to allow qvar = "" for
+  # elementary indexes
+  namesVector <- namesVector[!namesVector == ""]
+
   badNames <- namesVector[!(namesVector %in% goodNames)]
 
   if(length(badNames >= 1)){
@@ -74,9 +80,9 @@ checkTypes <- function(x, pvar, qvar, pervar){
 
   if(!inherits(x[[pervar]], "numeric")){
     coerced <- try(as.numeric(x[[pervar]]), silent = TRUE)
-    if(inherits(coerced, "try-error")){
+    if(inherits(coerced, "try-error") | any(is.na(coerced))){
       check = FALSE
-      message("Time period variable is not numeric and cannot be coerced to numeric")
+      m <- "Time period variable is not numeric and cannot be coerced to numeric"
     }
     else {
       x[[pervar]] <- coerced
@@ -85,23 +91,26 @@ checkTypes <- function(x, pvar, qvar, pervar){
 
   if(!inherits(x[[pvar]], "numeric")){
     coerced <- try(as.numeric(x[[pvar]]), silent = TRUE)
-    if(inherits(coerced, "try-error")){
+    if(inherits(coerced, "try-error") | any(is.na(coerced))){
       check = FALSE
-      message("Price variable is not numeric and cannot be coerced to numeric")
+      m <- "Price variable is not numeric and cannot be coerced to numeric"
     }
     else {
       x[[pvar]] <- coerced
     }
   }
 
-  if(!inherits(x[[qvar]], "numeric")){
-    coerced <- try(as.numeric(x[[qvar]]), silent = TRUE)
-    if(inherits(coerced, "try-error")){
-      check = FALSE
-      message("Quantity variable is not numeric and cannot be coerced to numeric")
-    }
-    else {
-      x[[qvar]] <- coerced
+  # don't check qvar for elementary indexes
+  if(!qvar == ""){
+    if(!inherits(x[[qvar]], "numeric")){
+      coerced <- try(as.numeric(x[[qvar]]), silent = TRUE)
+      if(inherits(coerced, "try-error") | any(is.na(coerced))){
+        check = FALSE
+        m <- "Quantity variable is not numeric and cannot be coerced to numeric"
+      }
+      else {
+        x[[qvar]] <- coerced
+      }
     }
   }
 
@@ -109,7 +118,7 @@ checkTypes <- function(x, pvar, qvar, pervar){
     return(x)
   }
   else {
-    stop("Please correct input data types")
+    stop(paste("Please correct input data types.", m), call. = FALSE)
   }
 
 }