+++ tests/testthat/test-dissimilarity.R
@@ -21,6 +21,11 @@ test_that("Dissimilarity measures return the correct values",{
   expect_equal(mixScaleDissimilarity(CES_sigma_2, pvar="prices",qvar="quantities",pervar="time",
                                      prodID = "prodID",measure = "scale", combine = "geomean"),
                testData$scale)
+  expect_equal(relativeDissimilarity(CES_sigma_2, pvar="prices",qvar="quantities",pervar="time",
+                                     prodID = "prodID", similarityMethod = "predictedshare"),
+               testData$predictedShare)
+  expect_equal(predictedShareDissimilarity(CES_sigma_2, pvar="prices", qvar="quantities", pervar="time",
+                                     prodID = "prodID"), testData$predictedShare)
 })
 
 test_that("maximum similarity links are estimated correctly",{
@@ -49,6 +54,9 @@ test_that("similarity linked bilateral price index functions return the correct 
                           prodID = "prodID",indexMethod = "laspeyres", sample="matched",
                           output = "chained", chainMethod = "mixscale", measure="scale")
                , testData$p_scale)
+  expect_equal(priceIndex(CES_sigma_2,pvar = "prices",qvar = "quantities",pervar = "time",
+                          prodID = "prodID",indexMethod = "laspeyres", sample="matched",
+                          output = "chained", chainMethod = "predictedshare"), testData$p_predictedShare)
 })
 
 rm(testData)